New version with reading routines from ifes TC
