+++ .gitignore
@@ -8,9 +8,13 @@
 logs/
 correlationHistogramAnalysis.egg-info/
 dist/
+correlationHistogramAnalysis/__pycache__/
 
 # Ignore IDE-specific files
 .idea/
 .vscode/
 .github/
+.venv/
+.pytest_cache/
+poetry.lock
 Shape_analysis.ipynb
