+++ data/final_energy_new.csv
@@ -1,0 +1,31 @@
+Molecule,E(0),E(1),E(2)
+Fe2O2,-2677.6173,-2677.3586,-2676.8874
+FeO2,-1414.023,-1413.6357,-1413.0053
+Fe2O,-2602.5007,-2602.2632,-2601.6958
+FeO3,-1489.1996,-1488.778,-1488.0879
+FeOH,-1339.4554,-1339.1836,-1338.5659
+FeH,-1264.1669,-1263.8948,-1263.2503
+HO2,-150.9123,-150.4981,-149.8103
+FeO4,-1564.287,-1563.8583,-1563.197
+OH,-75.7349,-75.2541,-74.492
+OH3,-76.8939,-76.6987,-75.8331
+H2O,-76.4261,-75.96723928,-75.04673655
+O,-75.0669,-74.55231762,-73.2584203
+Fe,-1263.5966,-1263.312824,-1262.703705
+O2,-150.3299,-149.873213,-149.0453982
+H,-0.498764285,NaN,NaN
+Fe2,-2527.196848,-2526.971348,-2526.489705
+FeO,-1338.818947,-1338.488558,-1337.840153
+H2,-1.173014918,-0.606016577,NaN
+Fe2O3,-2752.974325,-2752.646964,-2752.079671
+FeO2H,-1414.686467,-1414.323324,-1413.690356
+Fe2O4,-2828.16669,-2827.809819,-2827.199455
+H3,-1.665145584,-1.340014017,-0.521541359
+FeO2H2,-1415.358891,-1415.028125,-1414.400317
+H2O3,-226.69757,-226.3258682,-225.6018375
+FeH2O,-1340.029858,-1339.802812,-1339.278009
+FeH2,-1264.769079,-1264.416918,-1263.779689
+HO3,-226.057005,-225.6461,-225.10695
+H2O2,-151.54471,-151.16188,-150.35101
+O3,-225.407627,-224.9454555,-224.122016
+FeO3H2,-1490.5196,-1490.14782,-1489.50757
