+++ notebooks/README.md
@@ -1,0 +1,15 @@
+The example notebooks show the analysis of the correlation histograms of the three oxides.
+
+The epos files are not provided. The epos files used in the publication can be downloaded from here [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10677562.svg)](https://doi.org/10.5281/zenodo.10677562) after embargo.
+
+The notebooks can be run after installation of the correlationHistogramAnalysis module.
+
+All the notebooks have the following structure.
+ 1. import necessary modules.
+ 2. Processing DFT energies to identify possible reactions.
+ 3. Proceesing APT data (epos files) to overlay identified reactions in the correlation histograms.
+
+ Notes:
+ The current version of the code uses a ad-hoc readepos routines.
+
+ The future version of the code will use the reading routines provided from IFES Technical Committee. And also expand the use case of the current module.
