+++ pyproject.toml
@@ -1,30 +1,25 @@
-[build-system]
-requires = ["setuptools", "setuptools-scm"]
-build-backend = "setuptools.build_meta"
-
-[project]
+[tool.poetry]
 name = "correlationHistogramAnalysis"
-version = "0.2.1"
-license = { file = "LICENSE.txt" }
-authors = [
-  { name="Shyam Katnagallu", email="s.katnagallu@mpie.de" },
-]
-description = "A package for correlation histogram analysis in APT"
+version = "0.3.0"
+description = "\"For in depth analysis of correlation histograms in APT\""
+authors = ["Shyam Katnagallu <s.katnagallu@mpie.de>"]
+license = "APACHE 2.0"
 readme = "README.md"
-requires-python = ">=3.8"
-classifiers = [
-    "Programming Language :: Python :: 3",
-    "Operating System :: OS Independent",
+packages = [
+    { include = "correlationHistogramAnalysis" },
 ]
-maintainers = [
-  {name = "Shyam Katnagallu", email = "s.katnagallu@mpie.de"}
-]
-[tool.pytest.ini_options]
-minversion = "6.0"
-addopts = "-ra -q"
-testpaths = [
-    "tests/",
-]
-[project.urls]
-Homepage = "https://github.com/skatnagallu/CorrelationHistogram"
-Issues = "https://github.com/skatnagallu/CorrelationHistogram/issues"
+
+[tool.poetry.dependencies]
+python = "^3.12"
+numpy = "^1.26.4"
+pandas = "^2.2.1"
+ifes-apt-tc-data-modeling = "^0.2"
+
+[tool.poetry.group.dev.dependencies]
+ipykernel = "^6.29.2"
+matplotlib = "^3.8.3"
+scipy = "^1.12.0"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
